Moved NshmpError to new error package
