+++ webapp/apps/js/GmmDistance.js
@@ -2,7 +2,7 @@
 
 import Gmm from './lib/Gmm.js';
 import D3LinePlot from './lib/D3LinePlot.js';
-import NshmpError from './lib/NshmpError.js';
+import NshmpError from './error/NshmpError.js';
 import Tools from './lib/Tools.js';
 
 /**
