+++ webapp/apps/js/d3/data/D3LineDataBuilder.js
@@ -3,7 +3,7 @@ import D3LineData from './D3LineData.js';
 import D3LineSubView from '../view/D3LineSubView.js';
 import D3LineOptions from '../options/D3LineOptions.js';
 import D3LineSeriesData from './D3LineSeriesData.js'; 
-import NshmpError from '../../lib/NshmpError.js';
+import NshmpError from '../../error/NshmpError.js';
 
 /**
  * @fileoverview Builder for D3LineData
