+++ webapp/apps/js/d3/view/D3BaseView.js
@@ -3,7 +3,7 @@ import D3BaseSubView from './D3BaseSubView.js';
 import D3BaseSubViewOptions from '../options/D3BaseSubViewOptions.js';
 import D3BaseViewBuilder from './D3BaseViewBuilder.js';
 import D3BaseViewOptions from '../options/D3BaseViewOptions.js';
-import NshmpError from '../../lib/NshmpError.js';
+import NshmpError from '../../error/NshmpError.js';
 
 /**
  * @fileoverview Create a base view for plots to reside. The view
