+++ webapp/apps/js/lib/Hazard.js
@@ -3,7 +3,7 @@
 import Footer from './Footer.js';
 import Header from './Header.js';
 import LeafletTestSitePicker from './LeafletTestSitePicker.js';
-import NshmpError from './NshmpError.js';
+import NshmpError from '../error/NshmpError.js';
 import Settings from './Settings.js';
 import Spinner from './Spinner.js';
 import Tools from './Tools.js';
