+++ webapp/apps/js/lib/HazardNew.js
@@ -3,7 +3,7 @@
 import Constraints from './Constraints.js';
 import Footer from './Footer.js';
 import Header from './Header.js';
-import NshmpError from './NshmpError.js';
+import NshmpError from '../error/NshmpError.js';
 import Spinner from './Spinner.js';
 import Tools from './Tools.js';
