+++ webapp/apps/js/lib/LeafletTestSitePicker.js
@@ -2,7 +2,7 @@
 
 import TestSiteView from './TestSiteView.js';
 import Tools from './Tools.js';
-import NshmpError from './NshmpError.js';
+import NshmpError from '../error/NshmpError.js';
 
 /**
 * @class LeafletTestSitePicker
